requirements.txt
wyoming==1.2.0
azure-cognitiveservices-speech==1.32.1
black==23.10.0
lxml==4.9.3
pycountry==22.3.5